data_extractor/requirements.txt
apache-airflow==2.9.1
dask[complete]==2024.5.0
xarray==2024.3.0
rasterio==1.3.10
geopandas==0.14.4
numpy==1.26.4
scipy==1.13.0
requests==2.32.3
